Use t-plus as a helper instead of a component
