--- addon/components/t-plus/template.hbs
@@ -1,1 +1,0 @@
-{{t text preprintWords=translation _plus=attrs}}
