--- app/components/t-plus/component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-osf/components/t-plus/component';
