+++ app/helpers/t-plus.js
@@ -1,0 +1,1 @@
+export { default, tPlus } from 'ember-osf/helpers/t-plus';
