--- tests/integration/components/t-plus/component-test.js
@@ -1,25 +1,0 @@
-import { moduleForComponent, test } from 'ember-qunit';
-import hbs from 'htmlbars-inline-precompile';
-
-moduleForComponent('t-plus', 'Integration | Component | t plus', {
-  integration: true
-});
-
-test('it renders', function(assert) {
-
-  // Set any properties with this.set('myProperty', 'value');
-  // Handle any actions with this.on('myAction', function(val) { ... });
-
-  this.render(hbs`{{t}}`);
-
-  assert.equal(this.$().text().trim(), '');
-
-  // Template block usage:
-  this.render(hbs`
-    {{#t-plus}}
-      template block text
-    {{/t-plus}}
-  `);
-
-  assert.equal(this.$().text().trim(), 'template block text');
-});
