+++ tests/unit/helpers/t-plus-test.js
@@ -1,0 +1,12 @@
+
+import { tPlus } from 'dummy/helpers/t-plus';
+import { module, test } from 'qunit';
+
+module('Unit | Helper | t plus');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = tPlus([42]);
+  assert.ok(result);
+});
+
